Add support for UX3410UA / UX410UAK (single fan)
In BIOS version >= 308 fans are handled differently. This change reflects that by checking periodically what the temperature is and then only adjusting the minimum fan speed table value for the first / CPU fan at EC0 0x548. The automatic fan mode of the BIOS is kept on, so when the module fails somehow, the fan speed should recover within a couple of seconds.